requirements.txt
streamlit
azure-ai-inference
python-dotenv
pillow
requests